# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Favorite Products API",
        "description": "API for managing favorite product lists",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/lists": {
            "get": {
                "tags": [
                    "Favorite Lists"
                ],
                "summary": "Get user's favorite lists",
                "description": "Retrieve all favorite lists for the authenticated user",
                "operationId": "getFavoriteLists",
                "responses": {
                    "200": {
                        "description": "List of favorite lists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FavoriteList"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favorite Lists"
                ],
                "summary": "Create a new favorite list",
                "description": "Create a new favorite list for the authenticated user",
                "operationId": "createFavoriteList",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My Favorites"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Favorite list created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FavoriteList"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/lists/{id}": {
            "get": {
                "tags": [
                    "Favorite Lists"
                ],
                "summary": "Get favorite list with products",
                "description": "Retrieve a specific favorite list with its products",
                "operationId": "getFavoriteList",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Favorite list ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite list with products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "list": {
                                            "$ref": "#/components/schemas/FavoriteList"
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "List not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Favorite Lists"
                ],
                "summary": "Update favorite list",
                "description": "Update the name of a specific favorite list",
                "operationId": "updateFavoriteList",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Favorite list ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Favorites"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Favorite list updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FavoriteList"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "List not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Favorite Lists"
                ],
                "summary": "Delete favorite list",
                "description": "Soft delete a specific favorite list",
                "operationId": "deleteFavoriteList",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Favorite list ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite list deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "List deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "List not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/lists/{listId}/products": {
            "post": {
                "tags": [
                    "Favorite List Products"
                ],
                "summary": "Add product to favorite list",
                "description": "Add a product to a specific favorite list",
                "operationId": "addProductToList",
                "parameters": [
                    {
                        "name": "listId",
                        "in": "path",
                        "description": "Favorite list ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sku"
                                ],
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "example": "ABC123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product added to list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FavoriteListProduct"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "List not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/lists/{listId}/products/{sku}": {
            "delete": {
                "tags": [
                    "Favorite List Products"
                ],
                "summary": "Remove product from favorite list",
                "description": "Remove a product from a specific favorite list",
                "operationId": "removeProductFromList",
                "parameters": [
                    {
                        "name": "listId",
                        "in": "path",
                        "description": "Favorite list ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sku",
                        "in": "path",
                        "description": "Product SKU",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product removed from list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product removed from list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "List or product not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "FavoriteList": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "My Favorites"
                    },
                    "is_default": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "FavoriteListProduct": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "favorite_list_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sku": {
                        "type": "string",
                        "example": "ABC123"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "sku": {
                        "type": "string",
                        "example": "ABC123"
                    },
                    "name": {
                        "type": "string",
                        "example": "Sample Product"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "description": {
                        "type": "string",
                        "example": "A great product"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Favorite List Products",
            "description": "Operations for managing products in favorite lists"
        },
        {
            "name": "Favorite Lists",
            "description": "Favorite Lists"
        }
    ]
}